Cad Models
Gyatt
I was in Ohio before I met you
I rizzed too much and that's an issue but I'm grimace shake
Gyatt
Tell your friends it was nice to rizz them
But I hope I never edge again
I know it breaks your Fanum
Taxing in Ohio and I'm still not sigma
Four years no Livvy
Now you're looking pretty on Adin Ross' twitchy
And iiiiii can't rizz
No iiiiii can't mew
So baby Gronk me closer
In the back skibidi toilet
That I know you can't afford
Kai Cenat tatted on my shoulder
Pull the gyatt right off the corner
From that Fanum that you taxed
From your roommate back in Ohio
We ain't ever not the Rizzler!
We ain't ever not the Rizzler!
We ain't ever not the Rizzler!
You
Rizz as good as the day I met you
I forget just why I edged you
I was insane
Slay
And play that Blink-180 mew song
That we gooned to death in Ohio, okay?
I know you love this gyatt
I moved to Ohio in a brain rot car and
Four years no griddy
Now I'm lookin Livvy and you're not skibidi
And iiiiii can't rizz
No iiiiii can't mew
So baby Gronk me closer
In the back skibidi toilet
That I know you can't afford
Kai Cenat tatted on my shoulder
Pull the gyatt right off the corner
From that Fanum that you taxed
From your roommate back in Ohio
We ain't ever not the Rizzler!
We ain't ever not the Rizzler!
We ain't ever not the Rizzler!
So baby Gronk me closer
In the back skibidi toilet
That I know you can't afford
Kai Cenat tatted on my shoulder
Pull the gyatt right off the corner
From that Fanum that you taxed
From your roommate back in Ohio
We ain't ever not the Rizzler!
We ain't ever not the Rizzler!
Oh we ain't ever not the Rizzler!
We ain't ever not the Rizzler!
Oh we ain't ever not the Rizzler!
We ain't ever not the Rizzler!
Oh we ain't ever not the Rizzler!
We ain't ever not the Rizzler!
Oh we ain't ever not the Rizzler!
We ain't ever not the Rizzler!
We ain't ever not the Rizzler!